Fix code formatting for the API and demo

+++ example/demo.c
@@ -1,6 +1,7 @@
+#include <pthread.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <pthread.h>
+
 #include "ci_lib.h"
 
 #define BASE_VAL 10000
@@ -11,9 +12,10 @@
 
 void interrupt_handler(long ic) {
   /* This print should only appear in the CI integrated build */
-  static long previous_ic=0;
-  printf("Compiler interrupt called with instruction count %ld\n", ic-previous_ic);
-  previous_ic=ic;
+  static long previous_ic = 0;
+  printf("Compiler interrupt called with instruction count %ld\n",
+         ic - previous_ic);
+  previous_ic = ic;
 }
 
 void pre_disable() {
@@ -24,26 +26,24 @@ void post_enable() {
   printf("This function is being called after ci_enable()!\n");
 }
 
-void increment(void *arg)
-{
+void increment(void *arg) {
   int i;
   int counter = 0;
-  int thr_no = (int)arg;
-  int iterations = BASE_VAL + (rand()%10);
+  int thr_no = (int)((size_t)arg);
+  int iterations = BASE_VAL + (rand() % 10);
   register_ci(10000, 10000, interrupt_handler);
 
-  for(i=0; i<iterations; i++) {
+  for (i = 0; i < iterations; i++) {
     counter += rand() % 10;
   }
   printf("Counter: %d\n", counter);
 }
 
-void decrement(void *arg)
-{
+void decrement(void *arg) {
   int i;
   int counter = 0;
-  int thr_no = (int)arg;
-  int iterations = BASE_VAL + (rand()%10);
+  int thr_no = (int)((size_t)arg);
+  int iterations = BASE_VAL + (rand() % 10);
   register_ci(10000, 10000, interrupt_handler);
 
   /* disable CI for the remaining code */
@@ -54,17 +54,18 @@ void decrement(void *arg)
   register_ci_disable_hook(pre_disable);
   register_ci_enable_hook(post_enable);
 
-  /* There must be as many enable calls after a fixed number of disable calls, to make ci work. More number of enable calls will not make a difference */
+  /* There must be as many enable calls after a fixed number of disable calls,
+   * to make ci work. More number of enable calls will not make a difference */
   ci_disable(); /* will make no difference since CI is already disabled */
-  ci_enable(); /* won't be enabled yet since there were two disable calls before this */
+  ci_enable(); /* won't be enabled yet as there were two disable calls before */
   ci_enable(); /* will enable CI again */
 #endif
 
 #ifdef CHECK_DEREGISTER // define this flag to check de-register functionality
-  deregister(); /* deregisters ci */
+  deregister();         /* deregisters ci */
 #endif
 
-  for(i=0; i<iterations; i++) {
+  for (i = 0; i < iterations; i++) {
     counter -= rand() % 10;
   }
   printf("Counter: %d\n", counter);
@@ -75,31 +76,31 @@ int main(int argc, char **argv) {
   /* register interrupt handler */
   register_ci(10000, 10000, interrupt_handler);
 
-  pthread_t t[MAX_THREADS-1];
+  pthread_t t[MAX_THREADS - 1];
   int num_threads = MAX_THREADS;
-  if(argc==2) {
+  if (argc == 2) {
     num_threads = atoi(argv[1]);
-    if(num_threads > MAX_THREADS) {
+    if (num_threads > MAX_THREADS) {
       printf("Maximum thread count exceeded. Aborting.\n");
       exit(1);
     }
   }
 
   printf("Starting %d increment threads\n", num_threads);
-  for(int i=0; i<(num_threads-1); i++) {
-    pthread_create(&t[i], NULL, (void* (*)(void*))increment, (void*)&i);
+  for (int i = 0; i < (num_threads - 1); i++) {
+    pthread_create(&t[i], NULL, (void *(*)(void *))increment, (void *)&i);
   }
 
-  for(int i=0; i<(num_threads-1); i++) {
+  for (int i = 0; i < (num_threads - 1); i++) {
     pthread_join(t[i], NULL);
   }
 
   printf("Starting %d decrement threads\n", num_threads);
-  for(int i=0; i<(num_threads-1); i++) {
-    pthread_create(&t[i], NULL, (void* (*)(void*))decrement, (void*)&i);
+  for (int i = 0; i < (num_threads - 1); i++) {
+    pthread_create(&t[i], NULL, (void *(*)(void *))decrement, (void *)&i);
   }
 
-  for(int i=0; i<(num_threads-1); i++) {
+  for (int i = 0; i < (num_threads - 1); i++) {
     pthread_join(t[i], NULL);
   }
